Refactor Section implementation to allow more logic in model
Previously we had an enum representing each section, each case having an
array of objects representing the cells as associated values. The
problem was that we had to treat each of those arrays individually in
the UI layer which meant that some logic for determining which array
belonged in which section (first or second section) had to be in the UI
layer. Now we have a 2 x N 2d array of CellConfig enum objects; the
first dimension being sections and the second one being rows per
section. This has the advantage that the UI can simply map an IndexPath
to a CellConfig and configure the cell based on the cell config.