Move responsibility for saving the object to the flow coordinator
I want the new device to be added to the list after the pop of the edit
view is complete (so that the user sees the animation). We do this by
passing the object to be saved to the flow coordinator, the coordinator
pops the view controller and then does the save.