Handle adding a second section and moving a row to that section
In the case that we have 2 discovered devices and no known devices and
we add one device to the database. The result should be that we have 2
sections and one device in each section. To achieve that we need to add
the section, add the row to it and delete one of the rows from the
previous section.